+++ src/main/java/ua/foxminded/pinchuk/javaspring/carrestservice/entity/Brand.java
@@ -11,23 +11,23 @@ public class Brand {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @Column(name = "name")
     private String name;
 
     public Brand() {
     }
 
-    public Brand(Integer id, String name) {
+    public Brand(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
