+++ src/main/java/ua/foxminded/pinchuk/javaspring/carrestservice/entity/User.java
@@ -12,7 +12,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id")
-    private Integer userId;
+    private Long userId;
 
     @Column(name = "email")
     private String email;
@@ -37,7 +37,7 @@ public class User {
     public User() {
     }
 
-    public User(Integer userId, String email, String password,
+    public User(Long userId, String email, String password,
                 String firstName, String lastName, Role userRole,
                 String phone) {
         this.userId = userId;
@@ -49,11 +49,11 @@ public class User {
         this.phone = phone;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
